updated querykb tool to latest version many changes I don't remember by now :D

+++ my source code/Libraries/lib/querykb/querykb/src/test/java/com/githhub/aaronbembenek/querykb/AbstractTestCount.java
@@ -2,6 +2,7 @@ package com.githhub.aaronbembenek.querykb;
 
 import java.io.IOException;
 import java.io.StringReader;
+import java.math.BigInteger;
 
 import com.githhub.aaronbembenek.querykb.parse.ParseException;
 import com.githhub.aaronbembenek.querykb.parse.Parser;
@@ -15,11 +16,11 @@ public abstract class AbstractTestCount {
 		this.data = data;
 	}
 	
-	protected long query(String query) throws IOException, ParseException {
+	protected BigInteger query(String query) throws IOException, ParseException {
 		KnowledgeBase kb = KnowledgeBase.fromDatalogFacts(new StringReader(data));
 		Tokenizer t = new Tokenizer(new StringReader(query));
 		return query(kb, Parser.parseQuery(t));
 	}
 
-	protected abstract long query(KnowledgeBase kb, Query q);
+	protected abstract BigInteger query(KnowledgeBase kb, Query q);
 }
