+++ my source code/Libraries/lib/querykb/querykb/src/test/java/com/githhub/aaronbembenek/querykb/TestCountNaive.java
@@ -1,5 +1,7 @@
 package com.githhub.aaronbembenek.querykb;
 
+import java.math.BigInteger;
+
 import org.junit.runner.RunWith;
 import org.junit.runners.Suite;
 
@@ -10,8 +12,8 @@ public class TestCountNaive {
 	public static class GroundQueries extends AbstractTestCountGroundQueries {
 
 		@Override
-		protected long query(KnowledgeBase kb, Query q) {
-			return kb.countNaive(q);
+		protected BigInteger query(KnowledgeBase kb, Query q) {
+			return BigInteger.valueOf(kb.countNaive(q));
 		}
 
 	}
@@ -19,8 +21,8 @@ public class TestCountNaive {
 	public static class NonGroundQueries extends AbstractTestCountNonGroundQueries {
 
 		@Override
-		protected long query(KnowledgeBase kb, Query q) {
-			return kb.countNaive(q);
+		protected BigInteger query(KnowledgeBase kb, Query q) {
+			return BigInteger.valueOf(kb.countNaive(q));
 		}
 
 	}
